fix(Select): resolve enter keyboard event and filterable options
