+++ src/select/select.tsx
@@ -34,6 +34,12 @@ export default defineComponent({
     const COMPONENT_NAME = usePrefixClass('select');
     const { globalConfig, t } = useConfig('select');
     const { popupVisible, inputValue, modelValue, value } = toRefs(props);
+    const [innerInputValue, setInputValue] = useDefaultValue(
+      inputValue,
+      props.defaultInputValue,
+      props.onInputChange,
+      'inputValue',
+    );
     const [orgValue, setOrgValue] = useVModel(value, modelValue, props.defaultValue, props.onChange);
     const selectPanelRef = ref(null);
     const selectInputRef = ref(null);
@@ -41,7 +47,11 @@ export default defineComponent({
       label: props.keys?.label || 'label',
       value: props.keys?.value || 'value',
     }));
-    const { options, optionsMap, optionsList, optionsCache } = useSelectOptions(props, keys);
+    const { options, optionsMap, optionsList, optionsCache, displayOptions } = useSelectOptions(
+      props,
+      keys,
+      innerInputValue,
+    );
 
     // 内部数据,格式化过的
     const innerValue = computed(() => {
@@ -77,12 +87,6 @@ export default defineComponent({
       });
     };
 
-    const [innerInputValue, setInputValue] = useDefaultValue(
-      inputValue,
-      props.defaultInputValue,
-      props.onInputChange,
-      'inputValue',
-    );
     const [innerPopupVisible, setInnerPopupVisible] = useDefaultValue(
       popupVisible,
       false,
@@ -157,7 +161,7 @@ export default defineComponent({
     // 键盘操作逻辑
     const hoverIndex = ref(-1);
     const handleKeyDown = (e: KeyboardEvent) => {
-      const optionsListLength = optionsList.value.length;
+      const optionsListLength = displayOptions.value.length;
       let newIndex = hoverIndex.value;
       switch (e.code) {
         case 'ArrowUp':
@@ -278,6 +282,7 @@ export default defineComponent({
       isCheckAll: isCheckAll.value,
       onCheckAllChange,
       getSelectedOptions,
+      displayOptions: displayOptions.value,
     }));
 
     provide(selectInjectKey, SelectProvide);
